+++ src/RaynMaker.Portfolio.Specs/TDK.fs
@@ -1,5 +1,5 @@
 ﻿[<AutoOpen>]
-module RaynMaker.Portfolio.Specs.Foundation
+module RaynMaker.Portfolio.Specs.TDK
     open NUnit.Framework.Constraints
 
     let equalList x =